Database cleanup complete:
1.  Remove signal and block tables
2.  change device.moduleIndex to port
3.  Change controllerModule.moduleIndex to address
4.  change routeEntry.turnoutID to deviceID
5.  change signalAspectCondition to signalAspect
6.  change signalAspect.signalID to deviceID
7.  change signalCondition.signalAspectConditionID to signalAspectID

All code references to blockID, signalID and turnoutID changed to deviceID

Update MultiModuleController to use pin 12 as a Status LED indicator where ON means the controller is connected to the server.  Off, the controller is disconnected.

Update MultiModuleController to use pin 13 as a lockout switch pin.  When turned on (pulling to ground) the modules connected to the controller are informed of the lockout condition.  The (yet to be added) InputModule will use this to "lockout" button presses, etc.